+++ package/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Changes in `DirTarget` and `FileTarget`
 
+### What's New
+
+- Add `dir` alias function for `directory` function
+
 ### Migrating from 0.4.X to 0.5.0
 
 ```diff